PlutusConformance.Common

Description

Plutus conformance test suite library.

Documentation

shownParseError :: Text

The default shown text when a parse error occurs. We don't want to show the detailed parse errors so that users of the test suite can produce the expected output more easily.

shownEvaluationFailure :: Text

The default shown text when evaluation fails.

parseTxt :: Text -> Either ParserErrorBundle (Program Name DefaultUni DefaultFun SourcePos)

The default parser to parse UPLC program inputs.

type UplcProg = Program Name DefaultUni DefaultFun ()

The input/output UPLC program type.

type UplcEvaluator = UplcProg -> Maybe UplcProg

The evaluator to be tested. It should either return a program if the evaluation is successful, or Nothing if not.

discoverTests

Arguments

:: UplcEvaluator
The evaluator to be tested.

-> (FilePath -> Bool)
A function that takes a test name and returns whether it should be labelled as ExpectedFailure.

-> FilePath
The directory to search for tests.

-> IO TestTree

Walk a file tree, making test groups for directories with subdirectories, and test cases for directories without.

expectedToProg :: Text -> Either Text UplcProg

Turn the expected file content in text to a UplcProg unless the expected result is a parse or evaluation error.

getTestedValue :: UplcEvaluator -> FilePath -> IO (Either Text UplcProg)

Get the tested value. The tested value is either the shown parse or evaluation error, or a UplcProg.

compareAlphaEq

Arguments

:: Either Text UplcProg
golden value

-> Either Text UplcProg
tested value

-> Maybe String
If two values are the same, it returns Nothing. If they are different, it returns an error that will be printed to the user.

The comparison function used for the golden test. This function checks alpha-equivalence of programs when the output is a program.

updateGoldenFile

Arguments

:: FilePath
the path to write the golden file to

-> Either Text UplcProg

-> IO ()

Update the golden file with the tested value. TODO abstract out for other tests.

evalUplcProg :: UplcEvaluator

Our evaluator for the Haskell UPLC tests is the CEK machine.

runUplcEvalTests

Arguments

:: UplcEvaluator
The action to run the input through for the tests.

-> (FilePath -> Bool)
A function that takes a test name and returns whether it should labelled as ExpectedFailure.

-> IO ()

Run the UPLC evaluation tests given an evaluator that evaluates UPLC programs.

agdaEvalUplcProgDebug :: UplcProg -> Either String UplcProg

For debugging failed UPLC evaluation tests (Agda implementation). Called by `test-utils`.